Time: 16 ms (9.65%), Space: 23.8 MB (49.86%) - LeetHub

+++ 0283-move-zeroes/0283-move-zeroes.cpp
@@ -15,10 +15,10 @@ class Solution {
 public:
     void moveZeroes(vector<int>& nums) {
         
-        int n = nums.size();
-        int lastIndx = 0;
+        short n = nums.size();
+        short lastIndx = 0;
         short numOfZeros=0;
-        for (int j = 0; j < n && lastIndx<(n-numOfZeros) ; j++)
+        for (short j = 0; j < n && lastIndx<(n-numOfZeros) ; j++)
         {
             if (nums[j] != 0)
                 swap(nums[lastIndx++], nums[j]);
@@ -26,7 +26,7 @@ public:
                 numOfZeros++;   
         }
         
-        for (int j = lastIndx; j < n; j++)
+        for (short j = lastIndx; j < n; j++)
         {
             nums[j] = 0;
         }
